Starting: Parallel SGD for MNIST and CIFAR-10
To be adapted using the setup from mingpt, i.e. using optimizer, model, etc.
