+++ src/rawdata/cifar-10-batches-py/readme.html
@@ -1,0 +1,1 @@
+<meta HTTP-EQUIV="REFRESH" content="0; url=http://www.cs.toronto.edu/~kriz/cifar.html">
